Restructure metadata files
